Commence Project Cleanup!
I based the code off of team 254's code, which was awfully written. So
right now the ugly little critters need to be killed. So I've already
gone through the process of removing the Constants class. Everything is
now in the RobotMap class and is now much simpler. Next on the black
list is ControlBoard...