Cleaned the Catcher
Made the (physical) catcher subsystem completely command-based by
converting from periodically checking button values (LabVIEW style) to
mapping directly from buttons to commands. I did notice that there is
already a "FrontPanelCommand" class. However, I didn't spend that much
time looking at it, and came to the premature conclusion that it was a
result of naive coding. The production of OI and thus the destruction of
ControlBoard has begun!